canvas-agents/requirements.txt
flask>=3.0.0
python-dotenv>=1.0.0
pydantic>=2.6.0
requests>=2.31.0
